♻️ absorb mint or counterfeit into mint in store; put flag first
