+++ packages/atom.io/src/internal/families/mint-in-store.ts
@@ -4,46 +4,69 @@ import { stringifyJson } from "atom.io/json"
 
 import type { ReadableFamily } from "../state-types"
 import type { Store } from "../store"
-import { COUNTERFEIT, mint } from "../store"
+
+export const FAMILY_MEMBER_TOKEN_TYPES = {
+	atom_family: `atom`,
+	molecule_family: `molecule`,
+	mutable_atom_family: `mutable_atom`,
+	readonly_held_selector_family: `readonly_held_selector`,
+	readonly_pure_selector_family: `readonly_pure_selector`,
+	writable_held_selector_family: `writable_held_selector`,
+	writable_pure_selector_family: `writable_pure_selector`,
+} as const
 
 export const MUST_CREATE: unique symbol = Symbol(`MUST_CREATE`)
+export const DO_NOT_CREATE: unique symbol = Symbol(`DO_NOT_CREATE`)
 
-export function mintInStore<T, K extends Canonical, Key extends K, E>(
+export function mintInStore<T, K extends Canonical, KK extends K, E>(
+	mustCreate: typeof DO_NOT_CREATE | typeof MUST_CREATE,
 	store: Store,
 	family: ReadableFamily<T, K, E>,
-	key: Key,
-	mustCreate?: typeof MUST_CREATE,
-): WritableToken<T, Key, E>
-export function mintInStore<T, K extends Canonical, Key extends K, E>(
+	key: KK,
+): WritableToken<T, KK, E>
+export function mintInStore<T, K extends Canonical, KK extends K, E>(
+	mustCreate: typeof DO_NOT_CREATE | typeof MUST_CREATE,
 	store: Store,
 	family: ReadableFamily<T, K, E>,
-	key: Key,
-	mustCreate?: typeof MUST_CREATE,
-): ReadableToken<T, Key, E>
-export function mintInStore<T, K extends Canonical, Key extends K, E>(
+	key: KK,
+): ReadableToken<T, KK, E>
+export function mintInStore<T, K extends Canonical, KK extends K, E>(
+	mustCreate: typeof DO_NOT_CREATE | typeof MUST_CREATE,
 	store: Store,
 	family: ReadableFamily<T, K, E>,
-	key: Key,
-	mustCreate?: typeof MUST_CREATE,
-): ReadableToken<T, Key, E> {
+	key: KK,
+): ReadableToken<T, KK, E> {
 	const stringKey = stringifyJson(key)
 	const molecule = store.molecules.get(stringKey)
 
 	const cannotCreate = !molecule && store.config.lifespan === `immortal`
 
 	if (cannotCreate) {
+		const { type: familyType, key: familyKey } = family
 		store.logger.warn(
 			`💣`,
 			`key`,
 			stringKey,
 			`was used to mint a counterfeit token for`,
-			family.type,
-			`"${family.key}"`,
+			familyType,
+			`"${familyKey}"`,
 		)
-		return mint(family, key, COUNTERFEIT)
+		const fullKey = `${familyKey}(${stringKey})`
+		const type = FAMILY_MEMBER_TOKEN_TYPES[familyType]
+		const stateToken: ReadableToken<T, KK, E> & { counterfeit: true } = {
+			counterfeit: true,
+			key: fullKey,
+			type,
+			family: {
+				key: familyKey,
+				subKey: stringKey,
+			},
+		}
+
+		return stateToken
 	}
 
-	let token: ReadableToken<T, Key, E>
+	let token: ReadableToken<T, KK, E>
 	if (mustCreate === MUST_CREATE) {
 		store.logger.info(
 			`👪`,
@@ -57,7 +80,19 @@ export function mintInStore<T, K extends Canonical, Key extends K, E>(
 			store.moleculeData.set(stringKey, family.key)
 		}
 	} else {
-		token = mint(family, key)
+		const { type: familyType, key: familyKey } = family
+		const fullKey = `${familyKey}(${stringKey})`
+		const type = FAMILY_MEMBER_TOKEN_TYPES[familyType]
+		const stateToken: ReadableToken<T, KK, E> = {
+			key: fullKey,
+			type,
+			family: {
+				key: familyKey,
+				subKey: stringKey,
+			},
+		}
+
+		return stateToken
 	}
 
 	return token
